Modifiqué algunos modelos, ahora estoy viendo como hacer lo de autenticacion

+++ musicApi2/Models/Rating/Dto/UpdateRatingDto.cs
@@ -2,7 +2,7 @@
 {
     public class UpdateRatingDto
     {
-        public int RatingValue { get; set; }
+        public double RatingValue { get; set; }
 
         public int ReleaseId { get; set; }
